Fixing ClassSize violation on CodeNarc's own source (extracting chaining tests into standalone class).
